requirements.txt
openai
python-dotenv
ruff
duckduckgo_search
unstructured[pdf]
streamlit
litellm